Move `ClaimIndex` into its own module
